requirements.txt
timm==0.6.7
torch==2.0.1
plyfile==0.8.1
diffusers==0.24.0
peft==0.6.2
torchvision
scipy
numpy
imageio[ffmpeg]
opencv-python
Pillow
open3d